requirements.txt
unstructured
unstructured[pdf]
streamlit
openai
fastmcp
python-dotenv
uvicorn
bilibili-api-python
aiohttp
autogen-agentchat
autogen-ext[openai]
autogen-core
requests
google-genai
yt-dlp
static-ffmpeg
google-api-python-client
duckduckgo-search
trafilatura
pandas
